requirements.txt
psycopg2-binary~=2.9.10
pandas~=2.2.3
apscheduler~=3.10.4
python-dotenv~=1.0.1
fastapi~=0.115.7
uvicorn[standard]
requests~=2.32.3
scipy~=1.13.1
statistics~=1.0.3.5
pyjwt~=2.9.0
sqlalchemy~=2.0.36
matplotlib~=3.9.2
ccxt~=4.4.42
numpy~=2.0.1
mplfinance~=0.12.10b0
backtesting~=0.6.1

# formatting
black==25.1.0
isort==6.0.1